[ui] Encode URL params for community nux request (#20941)
## Summary & Motivation

We aren't currently encoding the values sent from the community NUX
dialog to our endpoint, which means that some email addresses are being
recorded incorrectly.

Repair this by using `URLSearchParams` to perform encoding on the vaues.

## How I Tested These Changes

Force community nux to appear in localhost. Submit the nux form with
various email addresses that need encoding, verify that they are
recorded correctly on the backend.