RFC: Expose Definitions (and RepositoryDefinition) on OpExecutionContext / AssetExecutionContext
Summary:
Provides a way to obtain the Definitions object that produced the job. Right now the only way to get this information for advanced use cases (for example, introspecting the asset graph in the body of an asset or op) is to either re-import it inline (potentially incurring a slow import again since it won't be cached by Python due to being imported in a different way) or rely on private APIs in the case of RepositoryDefinition

Exposed RepositoryDefinition as well since the above issues still apply for users who may have not fully migrated to Definitions yet.

Looking for initial high-level feedback, if this seems like a reasonable path forward I will flesh out / add full suite of tests / etc.