Add more granular run failure reason tracking when a step fails
Summary:
Before, we were assuming that if an individual step failed for any reason, it must have been due to a user code error. But this is not the case - if a step fails due to e.g. a database write failing or some other failure that is not in user code, we should treat that as a framework failure for the purpose of retrying.

Test Plan: BK
locally, set up an exception in a step subprocess but not in the run worker - set up run retries with retry_on_op_or_asset_failures = False. Run now retries, before it did not.