Freshness revamp (#20722)
Revamps freshness checks APIs.
- Use the same code path for the implementation, making the top level
functions one liners.
- Rename `maximum_lag_minutes` -> `lower_bound_delta`.
- Rename `freshness_cron` -> `deadline_cron`.
- Rename non-partitioned and time window partitioned builders to
`last_update` and `time_partition` builders, respectively.

The key idea here is that if we consider all assets to have a time
partition component, then we can unify the definition of freshness under
the hood. An asset is considered fresh if it's "latest relevant time
partition" was computed within the time window defined by some lower
bound and the current time. To break it into components:
1. The existence of the latest relevant time partition
2. The latest record representing this partition being from after some
lower bound

In `build_last_update_freshness_checks`, we consider the "latest
relevant time partition" to be any time partition (which trivializes to
no partition in the un-time-partitioned case), which means we only care
about (2).

In `build_time_partition_freshness_checks`, we consider the "latest
relevant time partition" to be the start of the time partition window
which has completed most recently from the last tick of the cron. There
is an implicit lower bound of the end of that time window which the
asset's update timestamp should always be greater than, in all cases,
meaning we only care about (1).

If we notice, the existence of these builders hinges on this idea of
defining what the "latest relevant time partition" is. We could imagine
a generalization here which allows you to define your own function for
determining this time partition, to handle more complex cases such as,
the current weekly partition having an update every day.

One thing worth noting is that the way we represent lower bound delta in
metadata is currently using seconds, which is probably undesirable in
many cases. Wondering if maybe makes sense to have a timedelta metadata
param? Pretty sure TimestampMetadataValue wouldn't render properly.

How I tested this
- Existing unit tests are sufficient. They are largely unchanged outside
of the two behavior changes called out earlier in this description, and
param renames.
- Added a new test for complex timezone interaction between the
execution timezone, the partitions timezone, and the freshness cron
timezone, all of which can technically be different.