test(dbt): expect no deps for column with ambiguous aggregate (#20937)
## Summary & Motivation
When using `*` in an aggregate function, `sqlglot` doesn't refine the
output to show that the resulting column is dependent on all of columns
found in the `FROM` statement. Instead, the resulting upstream lineage
for that column is empty.

So doing something like `count(*) as num_customers from {{
ref("customers") }}` would give an empty column lineage for
`num_customers`.

For now, I agree with this assessment, as the use of `*` is ambiguous
here and inferring that the lineage of column is "dependent" on all of
the columns feels wrong as it is potentially superfluous information. In
this PR, I am writing an explicit test case for it to ensure that
everyone else is on the same page with this behavior.

We could perhaps emit a warning in this case. We could recommend that
the user uses an explicit column name in their aggregation (e.g
`count(user_id)`) to ensure that the lineage flows properly.

## How I Tested These Changes
pytest