bugfix: Asset check unique_ids were not unique (#20949)
Now that checks are AssetsDefinitions, this hash needs to use them as
inputs. Are there any adverse implications to changing the hash?

Caught because this broke the subsetting logic with checks. Prior to the
refactor, we gave AssetChecksDefinitions `atomic_execution_unit_id =
None`. (now known as execution_set_identifier).
https://github.com/dagster-io/dagster/blob/release-1.6.0/python_modules/dagster/dagster/_core/host_representation/external_data.py#L1496
The refactor started grouping all asset checks in to one execution set
with the hash value for `[]`.

Added the test coverage to catch this (we weren't testing execution sets
of multiple `@asset_check`s).