[ui] Add another link from Events page to Partition > “View all events” modal (#20978)
## Summary & Motivation

Related: https://github.com/dagster-io/dagster/issues/20421

This adds another link to the Partition tab's "View Historical Events"
button to the individual event view. I think we need to re-think the
split between the partition-centric and event-centric pages and bring
partition filtering to the Events page, but until we're able to do that,
we need to make the existing modal easier to find.

Curious to get @braunjj's thoughts. We could also try to add a partition
filter to the events page, but I think right now we'd be limited to
choosing a single partition key.


![image](https://github.com/dagster-io/dagster/assets/1037212/c38b9b3d-b5e5-4b39-a675-21820aec5e57)

Co-authored-by: bengotow <bgotow@elementl.com>