[ui] Include assets with failed/missing partitions in “Materialize unsynced" (#20985)
Fixes #20463

## Summary & Motivation

This PR updates the behavior of "Materialize unsynced". In addition to
assets with the staleStatus = MISSING, we now include assets with
missing or failed partitions.

I also adjusted the types a bit in these code paths -- because
`staleCauses` can be undefined in the types passed to `isAssetStale`,
you could technically call it with "include=self" and never get anything
but false, which felt a little scary. I broke the function apart into
two with stricter types.

## How I Tested These Changes

Tested with a graph that includes some partially materialized
partitioned assets!

---------

Co-authored-by: bengotow <bgotow@elementl.com>