Add freshness_checks_toys (#20986)
Adds a source and derived asset freshness check. The source asset check
is scheduled alongside the observation, while the derived asset check
uses the build_sensor API to execute independently

Didn't add a new test per se but it's under test that the definitions
are buildable. Using a local build I tested that the unreliability
behavior works as expected, and that all states are possible.