feat(dbt): add `dagster-dbt project prepare-for-deployment` (#21016)
## Summary & Motivation
Introduces `dagster-dbt project prepare-for-deployment`. A one-stop shop for creating a manifest from an existing dbt project.

The idea here is that a user will invoke this command in CI/CD to produce all the manifests at build time.

## How I Tested These Changes
pytest: add tests where the project is prepared in different scenarios (e.g. to populate the package data dir, to populate state).