Also pin uv==1.29
It looks like uv also released around the same time. They're already on
top of getting 1.31 out the door but let's throw a pin in anyway.

We'll need to decide if this is something we want to manually increment
or if we're comfortable letting it always take latest. I also only added
the pin here but perhaps we need it everywhere we use uv (in all of our
tox.ini files, etc.)