Only create the specific GrapheneAssetNodes that you need when fetching an asset node with a given key
Summary:
Right now we construct a GrapheneAssetNode for every asset, then filter it down to the specific asset that we are looking for. Instead, pass the list of asset keys all the way through.

This isn't a substantial perf improvement yet, but future improvements that make the GrapheneAssetNode do less work when it is constructed would require this to be in place.