Exclude AllPartitionMappings from stale status calclation
Summary:
This adds another early breaking point based on profiling that we have observed causing unacceptably slow staleness calculations - an unpartitioned asset that is downstream of an AllPartitionsMapping. In this case, just calculating the list of partition keys can be enough to slow down the calculations enough here, and the thinking was that in the majority of cases there will be more than 100 partitions (and even if there were not, it would be odd and surprising for staleness calculations to suddenly disappear once the size of the partition set dips above 100 partitions).

I'm confused because there are comments here (and discussion in https://github.com/dagster-io/dagster/pull/14265/files) that seems to indicate we were already doing this, but I can't find any code that actually does it. Did we take it out at some point?