fix window on time partition freshness checks, description improvements (#21138)
We were previously starting the update time window at the deadline for
time partitioned assets, which is incorrect. The deadline is actually
the end of the partition time window.

We didn't catch this because there were no tests at the cusp of the
deadline, where it becomes clear that the update window doesn't make
sense. I added new tests which do cover this on-the-cusp case.

Also adds context to the description in the case that the asset hasn't
been updated within the time window at all (for the non partitioned
case). For the partitioned case, since we search for the specific
partition key, the observation is usually missing if the asset is out of
date, so i don't think we should consider that an unknown state.

Finally, the time descriptors I was using always used plural units,
changed to use singular if there's just one (it was annoying me).