Make the asset status cache do no additional DB work when the cursor is up to date and there are no in-progress runs materializing the asset
Summary:
The goal of this PR is to speed up the asset partition status cache in the hopefully reasonably common case where the cache is up to date and there are no in-progress runs currently materializing the asset. By leveraging the "last_planned_materialization_storage_id" field on AssetEntry, which is set in some storages but not others, we can add additional checks to short-circuit any DB queries using only the information that has already been fetched on the AssetEntry.