Fetch staleStatus separately in AssetGraphLiveQuery (#21205)
## Summary & Motivation

Splits out the fetching of staleStatus by breaking AssetLiveDataProvider
into two: AssetBaseDataProvider and AssetStaleDataProvider. The
AssetLiveDataProvider hook uses both and combines the result. For call
sites where we don't need the staleStatus we can use `useAssetBaseData`
instead of `useAssetLiveData`. Updated a couple of spots to use only the
base data.

One weird thing about this approach is that the staleData and baseData
would be fetched at different times so it's possible they are not
consistent with each other in some respect, I don't think there's anyway
to completely solve the problem because they're fetched in independent
requests so that could always happen but there might be something we
could do to lessen the probability of it happening, I'll take a stab at
this in a follow-up.

## How I Tested These Changes
Existing jest tests for AssetLiveDataProvider still pass.

Navigated around the app and made sure things still load, also paid
attention to network tab.

shadow dagit