:recycle: rework how `UPathIOManager` works with partitions (#21135)
## Summary & Motivation

This PR is addressing issues discussed in
https://github.com/dagster-io/dagster/pull/19294

The main addition is `UPathIOManager.load_partitions` method, which
should be considered a public interface for child IOManagers to
override. This interface is suited for `IOManager`s dealing with
serialization formats which support partitioning out of the box.

The benefits of this new interface are shown in `dagster-polars`, where
a lot of boilerplate and work-around code has been removed, and new
features are now supported thanks to `load_partitions`.

:sparkles: Additions
- Added `UPathIOManager.load_partitions` method. It's supposed to be
overridden by child classes which implement special handling for loading
multiple partitions. The default implementation loads them as a dict in
a `for` loop (just like before). Child classes may want to change this.
For example, for `DataFrame`s stored as `Parquet` datasets or
`DeltaLake`, it makes sense to use `partition_filters` instead, thus
loading multiple partitions in parallel.
- `dagster-polars`: now supports other data types rather than only
string for the partitioning columns
- `dagster-polars`: `PolarsDeltaIOManager` now supports
`MultiPartitionsDefinition` with `DeltaLake` native partitioning.
Metadata value `"partition_by": {"dim_1": "col_1", "dim_2": "col_2"}`
should be specified to enable this feature.

:heavy_minus_sign: Removed
- removed the type check which forced the result of multiple partitions
load to be a dictionary. Again, a single `DataFrame` can be loaded from
multiple partitions, and the return object won't be a dict.
 - removed a test which checked relevant behavior

:bug: Bug fixes
- fixed a bug with `MultipartitionKey`'s backcompat_path in
`UPathIOManager` - it was not used when loading multiple partitions
- `dagster-polars`: Previously, partition filtering happened after
reading the whole `DeltaLake` table. This has been fixed.

:warning: Breaking changes
- `PolarsDeltaIOManager` no longer supports loading natively partitioned
DeltaLake tables as dictionaries. They should be loaded as a single
`pl.DataFrame`/`pl.LazyFrame` instead.


## How I Tested These Changes
 - [x] existing tests
 - [x] using the new interface in `dagster-polars`
 - [x] added `MultiPartitionsDefinition` tests in `dagster-polars` 


Please let me know if you think this interface makes sense.