Docs/api reference css improvements (#21145)
## Summary & Motivation
These are a set of style changes.

### Improved Readability
Some developers had complaints about the overall readability of
Dagster's API reference. I've made changes to the global.css file to
improve the readability of the docs.

#### global.css changes
- **Line Height to 1.75em**: This makes it easier for the reader's eye
to move through the text which is helpful for information dense material
such as technical docs where certain passages may require multiple
readings to parse.
- **Parameters bolded**: The parameters in the headlines of class and
method docs are now bolded in order to offset them from their default
values. This makes it easier for the reader to quickly find params in
the list.

### Responsive 3 Panel Layout 
The Responsive 3 Panel layout is a common pattern for developer docs
sites. It uses responsive design to display 3 panels to viewers on large
viewports, two panels to users on medium viewports, and a single panel
on small viewports to give the user the best experience reading the docs
regardless of the size of their viewport.

It looks like an attempt was made to set this up for the dagster API
reference in the past, but something went wrong with the implementation.
The following actions resolved most of the issues.

#### Updating the right sidebar component in [...Pages].tsx
As it turned out, most of the 3 panel layout was implemented by a prior
developer. They simply forgot to apply the hidden tailwind class to the
right sidebar for viewports smaller than 1280px. This resulted in the
sidebar never closing, taking up a constant 380px regardless of
viewport, and generally messing up the way the rest of the CSS worked.

- **Added `hidden` to sidebar component**: The sidebar will vanish when
the viewport is smaller than 1280px. This also resolves another number
of display issues, including several odd overflows of breadcrumb and
header elements at diminished viewport size.

#### Updated global.css
- **Added overflow-wrap: word-break**: To all `<dl>` child elements in
the API reference to resolve overflow issues that showed up on mobile
devices.

## How I Tested These Changes
These are mostly style changes. Tested by viewing pages in the docs at
different viewport sizes including mobile emulation for iPhone SE.
Screenshots below.

### Desktop 3 panel View: 

![image](https://github.com/dagster-io/dagster/assets/39569270/b69c27ff-7eb5-43a1-a3cb-cff15d0934df)

### Desktop 2 Panel View

![image](https://github.com/dagster-io/dagster/assets/39569270/16f11580-dc43-430f-a25e-7096113baf8d)

### Desktop 1 Panel View

![image](https://github.com/dagster-io/dagster/assets/39569270/0a81b91d-4cce-4648-9f0b-b25d3c3eca81)

### Mobile

![image](https://github.com/dagster-io/dagster/assets/39569270/fa20f87b-a958-4d7c-bf9e-1a891de23180)