Docs/api reference css improvements (#21145)
These are a set of style changes.

Some developers had complaints about the overall readability of
Dagster's API reference. I've made changes to the global.css file to
improve the readability of the docs.

- **Line Height to 1.75em**: This makes it easier for the reader's eye
to move through the text which is helpful for information dense material
such as technical docs where certain passages may require multiple
readings to parse.
- **Parameters bolded**: The parameters in the headlines of class and
method docs are now bolded in order to offset them from their default
values. This makes it easier for the reader to quickly find params in
the list.

The Responsive 3 Panel layout is a common pattern for developer docs
sites. It uses responsive design to display 3 panels to viewers on large
viewports, two panels to users on medium viewports, and a single panel
on small viewports to give the user the best experience reading the docs
regardless of the size of their viewport.

It looks like an attempt was made to set this up for the dagster API
reference in the past, but something went wrong with the implementation.
The following actions resolved most of the issues.

As it turned out, most of the 3 panel layout was implemented by a prior
developer. They simply forgot to apply the hidden tailwind class to the
right sidebar for viewports smaller than 1280px. This resulted in the
sidebar never closing, taking up a constant 380px regardless of
viewport, and generally messing up the way the rest of the CSS worked.

- **Added `hidden` to sidebar component**: The sidebar will vanish when
the viewport is smaller than 1280px. This also resolves another number
of display issues, including several odd overflows of breadcrumb and
header elements at diminished viewport size.

- **Added overflow-wrap: word-break**: To all `<dl>` child elements in
the API reference to resolve overflow issues that showed up on mobile
devices.

These are mostly style changes. Tested by viewing pages in the docs at
different viewport sizes including mobile emulation for iPhone SE.
Screenshots below.

![image](https://github.com/dagster-io/dagster/assets/39569270/b69c27ff-7eb5-43a1-a3cb-cff15d0934df)

![image](https://github.com/dagster-io/dagster/assets/39569270/16f11580-dc43-430f-a25e-7096113baf8d)

![image](https://github.com/dagster-io/dagster/assets/39569270/0a81b91d-4cce-4648-9f0b-b25d3c3eca81)

![image](https://github.com/dagster-io/dagster/assets/39569270/fa20f87b-a958-4d7c-bf9e-1a891de23180)

 interactive rebase in progress; onto 79eacdfbef
 Last commands done (20 commands done):
    pick 4a2719439f [docs] tags and metadata change tracking docs (#21218)
    pick 0feef746ab Docs/api reference css improvements (#21145)
 Next commands to do (173 remaining commands):
    pick cd3ed18dac [docs] - Column-level lineage (DOC-157) (#21223)
    pick b20dd00f2f Update dbt tutorial with dbt command note (#21226)
 You are currently rebasing branch 'docs/mdx-component-to-markdoc' on '79eacdfbef'.

 Changes to be committed:
	modified:   docs/next/pages/[...page].tsx
	modified:   docs/next/styles/globals.css