Correctly handle deleted runs when determining in progress runs for the asset status partition cache
Summary:
If a run that planned to materialize an asset is deleted partway through, it currently stays marked as "in progress", since it is not returned in the query that checks for finished runs. Instead, remove runs that no longer exist from both the "in-progress" list and the "failed" list.