remove __del__ from event storage and poller (#20464)
`__del__` is bad vibes. 

This was originally added long ago before we got disciplined about using
context managers for handling the life cycle of `DagsterInstance`. That
should ensure proper shutdown in the vast majority of important
callsites.

Additional I have updated the `event_watcher` set-up to lazy instantiate
and clear itself on disposal. This further reduces the conditions where
`__del__` might have helped to only callsites that dont use a context
manager but are subscribing to the event stream for a run using this
non-public APIs.

## How I Tested These Changes

updated postgres and mysql tests