[ui] Repair run bucketing bug related to sort order (#21273)
## Summary & Motivation

Repair a batching bug in the run timeline.

The list of runs is sorted by start time (via the `left` pixel value)
prior to batching, but the current (descending) order of this value can
lead to incorrect batching, where batches that should overlap are not
bucketed together. The result is that rendered chunks can end up
visually sitting on top of each other.

Flip the sort order to fix the batching, and add a pixel of breathing
room on the "now" line to repair batching adjacent to that side.

## How I Tested These Changes

Added a Jest test for a case that passes with the ascending sort and
fails with descending sort.

View timeline with some of the known broken batching, verify that it is
now correct.