[ui] Use CSS vars for theming (#21200)
## Summary & Motivation

Use CSS vars instead of JS strings to set themed values.

The main benefit of this is that it will no longer require reloading the
page to see the theme update: when you change the theme in User
Settings, it will update immediately. Additionally, if the user's OS is
set to change from light to dark (and vice versa) based on time of day,
and they are using the "System" theme, the theme change should now occur
automatically, without a browser reload.

I have addressed a couple of items with inline comments.


https://github.com/dagster-io/dagster/assets/2823852/8e51bbce-6675-470b-b052-db7c51773bc4

## How I Tested These Changes

Load app, change theme. Verify that it updates right away. Click around
the app, verify that everything looks correct.