Fix deadline metadata (#21251)
The existing overdue_deadline_timestamp piece of metadata is confusing
and inscrutable. I'm replacing it by attempting to give the user the
information that they might want upon either a successful or failed
check result.

In the case of a successful check result, when's the next time that the
asset could potentially be not fresh / when do I next expect the asset?
In the case of a failed check, when did I expect the asset to arrive
initially?