[ui] Add context menu to non-global asset graphs, direction option to lineage (#21329)
## Summary & Motivation

Fixes FE-210 - previously the "background-click" context menu was only
available if groups were present in the graph. It's now available in the
single-group asset graph and the lineage asset graph as well.

In the lineage graph, you couldn't change the orientation OR
expand/collapse groups. I added the horizontal / vertical option so that
the context menu has something to display. We'll need to add this to the
column lineage graph as well, but I'll do that in a follow-up in cloud
since it'll re-use helpers in this PR.

## How I Tested These Changes

- Right click backgrounds of global, group, and lineage graphs
- Verify new horizontal vs. vertical toggle works as expected in the
lineage graph

Co-authored-by: bengotow <bgotow@elementl.com>