Make freshness check sensor work for all freshness checks (#21295)
Using the improved metadata from the previous PR, make it so that any
freshness check can be passed into build_freshness_check_sensor.

We greatly simplify the decisions the sensor makes. Essentially:
- If the previous check result was a success and has expired, check
again.
- Otherwise, don't check again.

We then only ever run checks when there is the possibility of state
change. Notice that if the check failed we won't ever run it again until
it succeeds (presumably when it runs right after a materialization or
observation), meaning a user won't get spammed with failures.

Revamped the test to cover all the cases that an asset check can be in
(success but expired, success and unexpired, never evaluated, failed
evaluation)

Since we're now using the DB to make decisions, and retrieving a row per
asset check key, I took special consideration for timeouts. We cursor on
the last unevaluated asset key once we reach 35 seconds of evaluation,
and then pick up from that point on the next eval.

Things that make me think this is a good change:
- Drastic decrease in LOC
- Provides guidelines for what a "freshness check" is. That is, any
check where, when it passes, we get fresh_until metadata.