[Fix] Cross-code-location partition mapping (#21353)
## Summary & Motivation

Fixes: https://github.com/dagster-io/dagster/issues/21306

The `asset_graph.get_partition_mapping()` method sneakily infers a
PartitionMapping if no explicit PartitionMapping is applied. This is
usually desirable, but when packaging up our external data, this means
that ALL dependencies will have a PartitionMapping defined (i.e. the
PartitionMapping will never be None). In many cases with
cross-code-location dependencies, users will not create a corresponding
SourceAsset object with the included PartitionsDefinition information,
and so the default inferred PartitionMapping between an (e.g.)
TimePartitioned asset and an (assumed-to-be-at-serialization-time)
unpartitioned dependency is the AllPartitionMapping.

This partition mapping then gets serialized, meaning that later on, when
we load multiple code locations worth of ExternalAssetNodes, we now have
an explicit partition mapping encoded, where one originally didn't
exist. This means that when we do the inference step again (now with the
true parent PartitionsDefinition), we are stuck with the
originally-serialized value of AllPartitionsMapping.

## How I Tested These Changes

Repro in the linked issue no longer repros, test failed before and
passes now.