Make `usePageContext` return the correct path no matter where its called (#21433)
## Summary & Motivation

To log page load we need to inform datadog when a new view starts. To do
this we currently rely on `usePageContext` to change when the path
changes. Currently it doesn't do that because `usePageContext` relies on
`useRouteMatch` from react-router-dom which finds the path from the
closest parent `Route`. The issue is that we're calling `usePageContext`
from a provider that is very high up in the tree, so `useRouteMatch`
can't access the context of the `<Route>` currently being rendered.
Instead of using `useRouteMatch` lets store the path in global state
when `useTrackPageView` is called and have `usePageContext` rely on that
global state.

Cloud requires this change to dedupe recoil:
https://github.com/dagster-io/internal/pull/9471

## How I Tested These Changes
👀 , play around with the app make sure nothing is broken