Remove add_input_metadata from UPathIOManager (#21423)
## Summary & Motivation

The current behavior of Dagster is that when `add_input_metadata` is
called on an input the in reality refers to an asset key, it generates
an AssetObservation for the asset.

I do think that behavior is suspect for a couple reasons since it
generates a ton of events and degrades perf, and I don't think
downstream assets should be generating asset observations for upstream
assets as a default behavior. The event stream for that asset will be
"polluted" by downstream assets that the user does not necessarily
control.

This PR does not re-examine that core behavior. However what it does to
is eliminate the "path" metadata from inputs for UPathIOManager. For
this IOManager the metadata is already encoded in the system, in the
corresponding upstream materialization. And this generates a ton of
low-value observation events in the default case.

## How I Tested These Changes

BK