Clean up text styles and simplify text style names (#21450)
## Summary & Motivation
This PR aims to clean up all of our text styles, and introduces a new
simplified naming convention that matches the style names within Figma.

The old text style naming convention of `body`, `body1`, and `body2` was
difficult to grok, ie: is body2 bigger or smaller than body1? When
should I `CaptionBold` or `SubtitleCaption`?

**Now we now have just 3 main text styles `Subtitle`, `Body`, and
`Mono`.**
- Each of these comes in a large and small variant, ie: `SubtitleLarge`|
`Subtitle` | `SubtitleSmall`
- There are 3 other less utilized styles `title`, `heading`, and `code`.

The full set of styles now looks like this, and IMO is much easier to
reason about.

![image](https://github.com/dagster-io/dagster/assets/2798333/dcc79be7-aa03-45a5-9d9b-81a57d6aae73)


## Legacy styles
I've renamed or removed all the previous styles except for `caption`
which is used in a lot of places. `caption` is the exact same style as
the new `bodySmall` style.

## In Figma
This has all been updated in Figma as well. In Dev mode, selecting any
text will now show it's corresponding style name in code.

![image](https://github.com/dagster-io/dagster/assets/2798333/4d34a002-1046-43be-a7f6-d08725e94e05)

## How I Tested These Changes
Loaded up the UI and clicked around.