Push down logic in is_sda_step into standalone function (#21472)
## Summary & Motivation

This is going to be a stack that batches and consolidates the logic to fetch upstream records. We are going to be building up a "preamble" to ExecutionStepContext construction that frontloads fetching. This is a first of series of PRs that does refactors to make that tractable. This just pushes down `is_sda_step` logic to a new standlone function `is_step_in_asset_graph_layer` that can be invoked before ExecutionStepContext is constructed

## How I Tested These Changes

BK