software-defined asset -> asset definition in docs (#21487)
## Summary & Motivation

This proposes using the term "asset definition" instead of
"software-defined asset" in docs. We would continue to use the word
"software-defined asset" in thought-leader-y content, to refer to the
idea of managing data assets using software.

Reasoning:
- All Dagster definitions come from software. Calling it out in this
particular case is inconsistent.
- "Software-defined asset" is a mouthful. This will matter even more in
a world where we treat what we now call "external assets" as a kind of
software-defined asset. Terms like "external software-defined asset" or
"materializable software-defined asset" are unnecessarily long and
difficult to work with.

## How I Tested These Changes